+++ tests/data/MainPageData.js
@@ -1,56 +1,56 @@
-export const content_test_page_root_child0000_child0010 = {":type": "test/components/componentchild0"};
-export const content_test_page_root_child0000_child0011 =  {":type": "test/components/componentchild1"};
+export const content_test_page_root_child0000_child0010 = {"cqType": "test/components/componentchild0"};
+export const content_test_page_root_child0000_child0011 =  {"cqType": "test/components/componentchild1"};
 
 export const content_test_page_root_child0000 = {
     "gridClassNames": "aem-Grid aem-Grid--12 aem-Grid--default--12",
     "columnCount": 12,
-    ":itemsOrder": ["child0010", "child0011"],
-    ":items": {
+    "cqItemsOrder": ["child0010", "child0011"],
+    "cqItems": {
         "child0010": content_test_page_root_child0000_child0010,
         "child0011":content_test_page_root_child0000_child0011
     },
-    ":type": "wcm/foundation/components/responsivegrid"
+    "cqType": "wcm/foundation/components/responsivegrid"
 };
 
-export const content_test_page_root_child0001 = {":type": "test/components/componentchild1"};
+export const content_test_page_root_child0001 = {"cqType": "test/components/componentchild1"};
 export const content_test_page_root = {
     "gridClassNames": "aem-Grid aem-Grid--12 aem-Grid--default--12",
     "columnCount": 12,
-    ":itemsOrder": ["child0000", "child0001"],
-    ":items": {
+    "cqItemsOrder": ["child0000", "child0001"],
+    "cqItems": {
         "child0000": content_test_page_root_child0000,
         "child0001":content_test_page_root_child0001
     },
-    ":type": "wcm/foundation/components/responsivegrid"
+    "cqType": "wcm/foundation/components/responsivegrid"
 };
 
 
-export const content_test_child_page_1_root_child1001 =  {":type": "test/components/componentchild1"};
+export const content_test_child_page_1_root_child1001 =  {"cqType": "test/components/componentchild1"};
 
 export const content_test_child_page_1_root_child1000 = {
     "gridClassNames": "aem-Grid aem-Grid--12 aem-Grid--default--12",
     "columnCount": 12,
-    ":type": "wcm/foundation/components/responsivegrid"
+    "cqType": "wcm/foundation/components/responsivegrid"
 };
 
 export const content_test_child_page_1_root = {
     "gridClassNames": "aem-Grid aem-Grid--12 aem-Grid--default--12",
     "columnCount": 12,
-    ":itemsOrder": ["child1000", "child1001"],
-    ":items": {
+    "cqItemsOrder": ["child1000", "child1001"],
+    "cqItems": {
         "child1000": content_test_child_page_1_root_child1000,
         "child1001": content_test_child_page_1_root_child1001
     },
-    ":type": "wcm/foundation/components/responsivegrid"
+    "cqType": "wcm/foundation/components/responsivegrid"
 };
 
 export const content_test_child_page_1 = {
-    ":type": "we-retail-journal/react/components/structure/page",
-    ":path": "/content/test/child_page_1",
-    ":items": {
+    "cqType": "we-retail-journal/react/components/structure/page",
+    "cqPath": "/content/test/child_page_1",
+    "cqItems": {
         "root": content_test_child_page_1_root
     },
-    ":itemsOrder": [
+    "cqItemsOrder": [
         "root"
     ]
 };
@@ -58,40 +58,40 @@ export const content_test_child_page_1 = {
 export const content_test_subpage2_root_child2000 = {
     "gridClassNames": "aem-Grid aem-Grid--12 aem-Grid--default--12",
     "columnCount": 12,
-    ":type": "wcm/foundation/components/responsivegrid"
+    "cqType": "wcm/foundation/components/responsivegrid"
 };
 
-export const content_test_subpage2_root_child2001 = {":type": "test/components/componentchild1"};
+export const content_test_subpage2_root_child2001 = {"cqType": "test/components/componentchild1"};
 
 export const content_test_subpage2_root = {
     "gridClassNames": "aem-Grid aem-Grid--12 aem-Grid--default--12",
     "columnCount": 12,
-    ":itemsOrder": ["child2000", "child2001"],
-    ":items": {
+    "cqItemsOrder": ["child2000", "child2001"],
+    "cqItems": {
         "child2000": content_test_subpage2_root_child2000,
         "child2001": content_test_subpage2_root_child2001
     },
-    ":type": "wcm/foundation/components/responsivegrid"
+    "cqType": "wcm/foundation/components/responsivegrid"
 };
 
 export const content_test_subpage2 = {
-    ":type":    "we-retail-journal/react/components/structure/page",
-    ":items": {
+    "cqType":    "we-retail-journal/react/components/structure/page",
+    "cqItems": {
         "root": content_test_subpage2_root
     },
-    ":itemsOrder": [
+    "cqItemsOrder": [
         "root"
     ]
 };
 
-export const content_test_subpage2_root_page =  {":type": "test/components/page"};
+export const content_test_subpage2_root_page =  {"cqType": "test/components/page"};
 
 export const content_test_subpage2_rootPage = {
-    ":type":    "we-retail-journal/react/components/structure/page",
-    ":items": {
+    "cqType":    "we-retail-journal/react/components/structure/page",
+    "cqItems": {
         "page": content_test_subpage2_root_page
     },
-    ":itemsOrder": [
+    "cqItemsOrder": [
         "page"
     ]
 };
@@ -103,19 +103,19 @@ export const PAGE_MODEL = {
     "templateName": "sample-template",
     "cssClassNames": "page",
     "language": "en-US",
-    ":itemsOrder": [
+    "cqItemsOrder": [
         "root"
     ],
-    ":items": {
+    "cqItems": {
         "root": content_test_page_root
     },
-    ":hierarchyType": "page",
-    ":children": {
+    "cqHierarchyType": "page",
+    "cqChildren": {
         "/content/test/child_page_1": content_test_child_page_1,
         "/content/test/subpage2/subpage22": content_test_subpage2_rootPage,
         "/content/test/subpage2": content_test_subpage2
     },
-    ":path": "/content/test/page",
-    ":type": "we-retail-react/components/structure/page"
+    "cqPath": "/content/test/page",
+    "cqType": "we-retail-react/components/structure/page"
 };