+++ tests/data/Page1Data.js
@@ -1,13 +1,13 @@
-export const content_test_page1_stem_child0000 = {":type": "test/components/componentchild1"};
+export const content_test_page1_stem_child0000 = {"cqType": "test/components/componentchild1"};
 
 export const content_test_page1_stem = {
     "gridClassNames": "aem-Grid aem-Grid--12 aem-Grid--default--12",
     "columnCount": 12,
-    ":itemsOrder": ["child0000"],
-    ":items": {
+    "cqItemsOrder": ["child0000"],
+    "cqItems": {
         "child0000": content_test_page1_stem_child0000
     },
-    ":type": "wcm/foundation/components/responsivegrid"
+    "cqType": "wcm/foundation/components/responsivegrid"
 };
 
 export const PAGE1 = {
@@ -17,12 +17,12 @@ export const PAGE1 = {
     "templateName": "sample-template",
     "cssClassNames": "page",
     "language": "en-US",
-    ":itemsOrder": [
+    "cqItemsOrder": [
         "stem"
     ],
-    ":items": {
+    "cqItems": {
         "stem": content_test_page1_stem
     },
-    ":path": "/content/test/page1",
-    ":type": "we-retail-react/components/structure/page"
+    "cqPath": "/content/test/page1",
+    "cqType": "we-retail-react/components/structure/page"
 };